pyproject.toml
[project]
name = "rag-lite"
version = "0.1.0"
description = "A lightweight RAG (Retrieval Augmented Generation) implementation"
authors = []
dependencies = [
  "python-dotenv>=1.0.0",    # .env loader
  "pydantic>=2.0.0",         # config & data validation
  "python-docx>=0.8.11",     # docx parsing
  "mammoth>=1.9.1",          # docx to markdown
  "streamlit>=1.25.0",       # UI
  "numpy>=1.24.4",           # numeric core
  "openai>=1.82.1",          # OpenAI API calls
  "boto3>=1.38.32",          # S3 cache
  "transformers>=4.52.4",    
  "sentence-transformers>=4.1.0",
  "faiss-cpu>=1.10.0",
  "html2text",
  "ollama>=0.5.1",
  "humanfriendly",
  "marker-pdf>=1.8.2",
]



requires-python = ">=3.11"

[project.optional-dependencies]
cloud = [
  "openai>=1.82.1",          # OpenAI API calls
  "boto3>=1.38.32",          # S3 cache
]
local = [
  "transformers>=4.52.4",    
  "sentence-transformers>=4.1.0",
  "faiss-cpu>=1.10.0",
]
dev = ["ruff", "pytest"]
docx = ["python-docx"]
pdf = ["PyMuPDF"]
pandoc = []  # reminder: Pandoc is a system dependency

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rag-lite"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
